initial Linux port
This gets GODS compiling on Linux with GCC, but we haven’t done any
testing to find out if any of the functions still work.  Unfortunately
(?) the Linux portability work was mixed with a lot of minor code
cleanups to get rid of warnings, which may not have been justifiable.